% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcompetitiveKineticEnzyme.R
\encoding{UTF-8}
\name{AcompetitiveKineticEnzyme}
\alias{AcompetitiveKineticEnzyme}
\title{AcompetitiveKineticEnzyme: Acompetitive Inhibition Fitting}
\usage{
AcompetitiveKineticEnzyme(sb, rate, inh)
}
\arguments{
\item{sb}{Substrate concentrations. A numeric vector representing the substrate concentrations used in the enzyme assay.}

\item{rate}{Reaction rates. A numeric vector representing the corresponding reaction rates observed at each substrate concentration.}

\item{inh}{Inhibitor concentrations. A numeric vector representing the inhibitor concentrations used in the enzyme assay.}
}
\value{
A data frame containing the estimated parameters and other analysis results. The data frame includes the following columns:
\item{AIC}{The Akaike Information Criterion (AIC) value for model selection.}
\item{BIC}{The Bayesian Information Criterion (BIC) value for model selection.}
\item{Km}{The estimated Michaelis-Menten constant.}
\item{Vm}{The estimated maximum reaction rate.}
\item{Kic}{The estimated inhibition constant.}
\item{Kiu}{The estimated uncompetitive inhibition constant.}
\item{StandardErrorvm}{The standard error associated with the estimated maximum reaction rate (Vm).}
\item{StandardErrorkm}{The standard error associated with the estimated Michaelis-Menten constant (Km).}
\item{StandardErrorkic}{The standard error associated with the estimated inhibition constant (Kic).}
\item{StandardErrorkiu}{The standard error associated with the estimated uncompetitive inhibition constant (Kiu).}
}
\description{
This function fits the acompetitive inhibition model to the given data. Acompetitive inhibition occurs when an inhibitor binds to both the enzyme and the enzyme-substrate complex, reducing the reaction rate. This function estimates the parameters of the acompetitive inhibition model using non-linear least squares fitting.
}
\examples{
f<-"https://www.ugr.es/~bioest/data/inhibicionnc.txt"
datos<-read.table(f,sep=",",header = T)
AcompetitiveKineticEnzyme(sb=datos$substrate,rate=datos$rate,inh=datos$inhibitor)

}
