% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompetitiveKineticEnzyme.R
\encoding{UTF-8}
\name{CompetitiveKineticEnzyme}
\alias{CompetitiveKineticEnzyme}
\title{CompetitiveKineticEnzyme: Competitive Inhibition Fitting}
\usage{
CompetitiveKineticEnzyme(sb, rate, inh)
}
\arguments{
\item{sb}{A numeric vector representing the substrate concentrations.}

\item{rate}{A numeric vector representing the corresponding reaction rates.}

\item{inh}{A numeric vector representing the concentrations of the competitive inhibitor.}
}
\value{
A data frame containing the estimated parameters and other analysis results. The data frame includes the following columns:
\item{AIC}{The Akaike Information Criterion (AIC) value for model selection.}
\item{BIC}{The Bayesian Information Criterion (BIC) value for model selection.}
\item{Km}{The estimated Michaelis-Menten constant.}
\item{Vm}{The estimated maximum reaction rate.}
\item{Kic}{The estimated inhibition constant.}
\item{Kiu}{The estimated uncompetitive inhibition constant.}
\item{StandardErrorvm}{The standard error associated with the estimated maximum reaction rate (Vm).}
\item{StandardErrorkm}{The standard error associated with the estimated Michaelis-Menten constant (Km).}
\item{StandardErrorkic}{The standard error associated with the estimated inhibition constant (Kic).}
\item{StandardErrorkiu}{The standard error associated with the estimated uncompetitive inhibition constant (Kiu).}
}
\description{
This function fits the Competitive Inhibition model to the given data using the least squares method. The Competitive Inhibition model describes the enzyme kinetics in the presence of a competitive inhibitor.
}
