% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InhibitionComparisonKineticEnzyme.R
\encoding{UTF-8}
\name{InhibitionComparisonKineticEnzyme}
\alias{InhibitionComparisonKineticEnzyme}
\title{InhibitionComparisonKineticEnzyme: Compare Different Enzyme Kinetic Inhibition Models}
\usage{
InhibitionComparisonKineticEnzyme(sb, rate, inh)
}
\arguments{
\item{sb}{A vector or column of substrate concentrations.}

\item{rate}{A vector or column of corresponding reaction rates.}

\item{inh}{A vector or column of inhibitor concentrations.}
}
\value{
A data frame containing the results of the three inhibition models.
The data frame includes the following columns:
\item{AIC}{Akaike Information Criterion for each model.}
\item{BIC}{Bayesian Information Criterion for each model.}
\item{Standard error for Vm}{Standard error of the maximum velocity parameter for each model.}
\item{Standard error for Km}{Standard error of the Michaelis constant parameter for each model.}
\item{Standard error for kic}{Standard error of the inhibitor constant parameter for each model.}
\item{Standard error for Kiu}{Standard error of the uncompetitive inhibitor constant parameter for each model.}
}
\description{
This function compares different enzyme kinetic inhibition models by fitting them to the given data.
The function requires the 'KineticEnzyme' package to be installed.
The function calculates the Akaike Information Criterion (AIC) and Bayesian Information Criterion (BIC)
for each model and provides additional analysis results such as the estimated parameters and their
corresponding standard errors.
The function also plots the data and the fitted curves for each model.
}
\examples{
f<-"https://www.ugr.es/~bioest/data/inhibicionnc.txt"
data<-read.table(f,sep=",",header = T)
InhibitionComparisonKineticEnzyme(sb=data$substrate,inh=data$inhibitor,rate=data$rate)

}
