% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KineticEnzymepackage.R
\docType{package}
\name{KineticEnzyme}
\alias{KineticEnzyme}
\title{KineticEnzyme Package}
\description{
This package provides a comprehensive solution for data analysis
related to the Michaelis Menten model, including linearization of
the data and consideration of the three main cases of enzyme
inhibition. In addition, estimates of great relevance in the
biochemical field are offered, such as the estimated parameters,
the standard errors and the residual analysis in those cases in
which it is necessary.
First, the package takes care of preparing the data, removing
the rows that contain zeros. Also, if specified, outliers can
be removed using Tukey's criterion. Next, we proceed to estimate
the model parameters using linear regression and calculate
the values predicted by the model.
If required, the package offers additional analysis and
diagnostics on model fit. This involves calculating the
correlation, performing model comparisons using criteria
such as ANOVA, AIC, and BIC, and analyzing the residuals
using tests for normality, homoscedasticity, and autocorrelation.
In addition, confidence intervals for the estimated
parameters can be calculated and the data plotted along
with the corresponding confidence bands, making it easy
to visually interpret the results obtained. In summary,
the R package provides a complete and cohesive tool for
the analysis of data related to the Michaelis Menten model,
providing a wide range of functionalities and facilitating interpretation and decision making in the biochemical field.
}
\author{
Alba Sevilla Navarro \email{sevillanavarro.alba@gmail.com}
}
