% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBKineticEnzyme.R
\encoding{UTF-8}
\name{LBKineticEnzyme}
\alias{LBKineticEnzyme}
\title{LBKineticEnzyme: Lineweaver-Burk Model Fitting}
\usage{
LBKineticEnzyme(substrate, velocity, removeoutliers = FALSE, deepening = FALSE)
}
\arguments{
\item{substrate}{Sustrate data. A numeric vector containing the substrate concentrations.}

\item{velocity}{Velocity data. A numeric vector containing the corresponding reaction velocities.}

\item{removeoutliers}{Logical value indicating whether to remove outliers from the data.
If set to TRUE, the function will remove outliers using a suitable method
(e.g., the Hampel identifier or the Tukey's fences method) before fitting the model.
Default is FALSE.}

\item{deepening}{Logical value indicating whether to perform in-depth analysis and diagnostics.
If set to TRUE, the function will provide additional diagnostic plots and statistics
to assess the goodness of fit and identify potential issues with the model.
Default is FALSE.}
}
\value{
A data frame containing the estimated parameters and other analysis results.
}
\description{
This function fits the Lineweaver-Burk model to the given data.
The Lineweaver-Burk model is a linearized form of the Michaelis-Menten equation
used to analyze enzyme kinetics.
}
