% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMKineticEnzyme.R
\encoding{UTF-8}
\name{MMKineticEnzyme}
\alias{MMKineticEnzyme}
\title{MMKineticEnzyme: Michaelis Menten Model Fitting}
\usage{
MMKineticEnzyme(substrate, velocity, removeoutliers = FALSE)
}
\arguments{
\item{substrate}{A vector or column of substrate concentrations.}

\item{velocity}{A vector or column of corresponding velocity measurements.}

\item{removeoutliers}{Logical value indicating whether to remove outliers from the data. Default is FALSE.}
}
\value{
A data frame containing the following columns:
\item{AIC}{Akaike Information Criterion.}
\item{BIC}{Bayesian Information Criterion.}
\item{vmax}{Estimated maximum velocity parameter.}
\item{km}{Estimated Michaelis-Menten constant parameter.}
\item{StandardErrorvm}{Standard error of the estimated maximum velocity parameter.}
\item{StandardErrorkm}{Standard error of the estimated Michaelis-Menten constant parameter.}

The function also plots the data and the fitted curve, and adds confidence bands to the plot.
}
\description{
This function fits the Michaelis Menten model to the given data.
It estimates the parameters of the model using the nlsLM function from the minpack.lm library.
The function also calculates the standard errors of the estimated parameters and provides
additional analysis results such as the Akaike Information Criterion (AIC) and Bayesian Information Criterion (BIC).
}
\examples{
s<-c(0.5,1.0,2.4,4.4,7.0,8.0,13.1,17.6,20.2,30.0)
v<-c(2.12,3.33,4.90,6.22,7.01,7.51,7.87,8.44,8.76,8.98)
data <- cbind(s,v)
data<-data.frame(data)
MMKineticEnzyme(substrate=data$s,velocity=data$v,removeoutliers=TRUE)
}
