% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonCompetitiveKineticEnzyme.R
\encoding{UTF-8}
\name{NonCompetitiveKineticEnzyme}
\alias{NonCompetitiveKineticEnzyme}
\title{NonCompetitiveKineticEnzyme: noncompetitive inhibition Fitting}
\usage{
NonCompetitiveKineticEnzyme(sb, rate, inh)
}
\arguments{
\item{sb}{Substrate data. A numeric vector containing the substrate concentrations.}

\item{rate}{Rate data. A numeric vector containing the corresponding reaction rates.}

\item{inh}{Inhibitor data. A numeric vector containing the inhibitor concentrations.}
}
\value{
A data frame containing the estimated parameters and other analysis results. The data frame includes the following columns:
\item{AIC}{The Akaike Information Criterion (AIC) value for model selection.}
\item{BIC}{The Bayesian Information Criterion (BIC) value for model selection.}
\item{Km}{The estimated Michaelis-Menten constant.}
\item{Vm}{The estimated maximum reaction rate.}
\item{kic}{The estimated inhibition constant.}
\item{StandardErrorvm}{The standard error associated with the estimated maximum reaction rate (Vm).}
\item{StandardErrorkm}{The standard error associated with the estimated Michaelis-Menten constant (Km).}
\item{StandardErrorkic}{The standard error associated with the estimated inhibition constant (kic).}
}
\description{
This function fits the noncompetitive inhibition model to the given data.
The noncompetitive inhibition model describes the relationship between substrate concentration,
reaction rate, and inhibitor concentration in enzyme kinetics.

The function takes substrate, rate, and inhibitor data as input and performs the model fitting.
It estimates the parameters Km, Vm, kic, using the nonlinear least squares method.
The function also provides a plot of the fitted lines and returns a data frame containing the
estimated parameters and other analysis results.
}
\examples{
f<-"https://www.ugr.es/~bioest/data/inhibicionnc.txt"
datos<-read.table(f,sep=",",header = T)
head(datos)
NonCompetitiveKineticEnzyme(sb=datos$substrate,rate=datos$rate,inh=datos$inhibitor)

}
