% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMMKineticEnzyme.R
\encoding{UTF-8}
\name{SimMMKineticEnzyme}
\alias{SimMMKineticEnzyme}
\title{SimMMKineticEnzyme: Simulate Samples for the M-M Equation}
\usage{
SimMMKineticEnzyme(
  n = 0,
  vm,
  km,
  srange = c(0, 1),
  sinc = 0.1,
  sigma = 0.06,
  het = TRUE
)
}
\arguments{
\item{n}{Number of points to generate. If a value greater than zero is specified,
a linear sequence of length "n" will be generated within the specified range.
If zero is specified, a sequence with an increment of "sinc" will be generated within the specified range.}

\item{vm}{Maximum velocity (Vmax) value in the M-M equation.}

\item{km}{Michaelis-Menten constant (Km) value in the M-M equation.}

\item{srange}{Range of substrate (S) values for which the samples will be generated.
Specified as a vector with the minimum and maximum values of the range.}

\item{sinc}{Increment value for the linear sequence of substrate values. Applicable only when "n" is zero.}

\item{sigma}{Standard deviation of the random noise added to the generated samples.}

\item{het}{Logical value indicating whether to introduce heteroscedasticity in the samples. Default is TRUE.}
}
\value{
A data frame with the generated samples, including the substrate (S) values, the observed velocity (v),
  and the model-predicted velocity (vmod) based on the M-M equation.
}
\description{
This function generates samples for the Monod-Michel (M-M) equation, which describes the
relationship between the concentration of a substrate and the rate of an enzymatic reaction.
}
\examples{
data<-SimMMKineticEnzyme(n=20,vm=1,km=0.3)
plot(data$s,data$v)
}
