% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAKineticEnzyme.R
\encoding{UTF-8}
\name{EAKineticEnzyme}
\alias{EAKineticEnzyme}
\title{EAKineticEnzyme: Eadie Scatchard Model Fitting}
\usage{
EAKineticEnzyme(substrate, velocity, removeoutliers = FALSE, deepening = FALSE)
}
\arguments{
\item{substrate}{Sustrate data. A numeric vector containing the substrate concentrations.}

\item{velocity}{Velocity data. A numeric vector containing the corresponding reaction velocities.}

\item{removeoutliers}{Logical value indicating whether to remove outliers from the data.
If set to TRUE, the function will remove outliers using a suitable method
(e.g., the Hampel identifier or the Tukey's fences method) before fitting the model.
Default is FALSE.}

\item{deepening}{Logical value indicating whether to perform in-depth analysis and diagnostics.
If set to TRUE, the function will provide additional diagnostic plots and statistics
to assess the goodness of fit and identify potential issues with the model.
Default is FALSE.}
}
\value{
A data frame containing the estimated parameters and other analysis results.
}
\description{
This function fits the Eadie Scatchard model to the given data
and provides a comprehensive analysis of the model. The Eadie Scatchard model
is a kinetic model used to analyze enzyme-substrate interactions and determine
important parameters such as the maximum velocity (Vmax) and the Michaelis constant (Km).
The function takes two input vectors, substrate and velocity,
representing the substrate concentrations and the corresponding
velocity measurements, respectively. It also includes optional
parameters: removeoutliers and deepening to control outlier
removal and in-depth analysis, respectively.
}
