% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelComparisonKineticEnzyme.R
\encoding{UTF-8}
\name{ModelComparisonKineticEnzyme}
\alias{ModelComparisonKineticEnzyme}
\title{ModelComparisonKineticEnzyme: Compare Different Enzyme Kinetic Models}
\usage{
ModelComparisonKineticEnzyme(substrate, velocity)
}
\arguments{
\item{substrate}{A vector or column of substrate concentrations.}

\item{velocity}{A vector or column of corresponding velocity measurements.}
}
\value{
A data frame containing the following columns:
\item{AIC}{Akaike Information Criterion for each model.}
\item{BIC}{Bayesian Information Criterion for each model.}
\item{vmax}{Estimated maximum velocity parameter for each model.}
\item{km}{Estimated Michaelis-Menten constant parameter for each model.}
\item{StandardErrorvm_delta}{Standard error of the estimated maximum velocity parameter for each model using delta method.}
\item{StandardErrorkm_delta}{Standard error of the estimated Michaelis-Menten constant parameter for each model using delta method.}

The row names of the data frame correspond to the names of the enzyme kinetic models.

Additionally, the function prints the following information:
\item{Enzyme kinetic model with the lowest AIC value}{The enzyme kinetic model with the lowest AIC value.}
\item{Enzyme kinetic model with the lowest BIC value}{The enzyme kinetic model with the lowest BIC value.}
\item{Enzyme kinetic model with the lowest Log Likelihood value}{The enzyme kinetic model with the lowest Log Likelihood value.}
\item{Enzyme kinetic model with the lowest standard error of the maximum velocity (vm) parameter}{The enzyme kinetic model with the lowest standard error of the maximum velocity (vm) parameter.}
\item{Enzyme kinetic model with the lowest standard error of the Michaelis-Menten (km) constant parameter}{The enzyme kinetic model with the lowest standard error of the Michaelis-Menten (km) constant parameter.}
}
\description{
This function compares different enzyme kinetic models by fitting them to the given data.
    It calculates the Akaike Information Criterion (AIC) and Bayesian Information Criterion (BIC)
    for each model, as well as the estimated parameters and their standard errors.
    The function supports the following enzyme kinetic models: Michaelis-Menten, Lineweaver-Burk,
    Eadie-Scatchard, Woolf-Augustinson-Hofstee, and Hanes-Woolf.
}
\examples{
s<-c(0.5,1.0,2.4,4.4,7.0,8.0,13.1,17.6,20.2,30.0)
v<-c(2.12,3.33,4.90,6.22,7.01,7.51,7.87,8.44,8.76,8.98)
data <- cbind(s,v)
data<-data.frame(data)
ModelComparisonKineticEnzyme(substrate=data$s,velocity=data$v)
}
