% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAHKineticEnzyme.R
\encoding{UTF-8}
\name{WAHKineticEnzyme}
\alias{WAHKineticEnzyme}
\title{WAHKineticEnzyme: Woolf Augustinson Hofstee Model Fitting}
\usage{
WAHKineticEnzyme(
  substrate,
  velocity,
  removeoutliers = FALSE,
  deepening = FALSE
)
}
\arguments{
\item{substrate}{Sustrate data. A numeric vector containing the substrate concentrations.}

\item{velocity}{Velocity data. A numeric vector containing the corresponding reaction velocities.}

\item{removeoutliers}{Logical value indicating whether to remove outliers from the data.
If set to TRUE, the function will remove outliers using a suitable method
(e.g., the Hampel identifier or the Tukey's fences method) before fitting the model.
Default is FALSE.}

\item{deepening}{Logical value indicating whether to perform in-depth analysis and diagnostics.
If set to TRUE, the function will provide additional diagnostic plots and statistics
to assess the goodness of fit and identify potential issues with the model.
Default is FALSE.}
}
\value{
A data frame containing the estimated parameters and other analysis results.
}
\description{
This function fits the Woolf Augustinson Hofstee model to the given data.
The model is a non-linear regression model commonly used in enzyme kinetics
to describe the relationship between substrate concentration and reaction velocity.
The model is given by V = (Vmax * [S]) / (Km + [S]), where V is the reaction velocity,
Vmax is the maximum velocity, [S] is the substrate concentration, and Km is the Michaelis-Menten constant.
The function takes substrate and velocity data as input and performs the model fitting.
It also provides options to remove outliers from the data and perform in-depth analysis and diagnostics.
}
\examples{
s<-c(0.5,1.0,2.4,4.4,7.0,8.0,13.1,17.6,20.2,30.0)
v<-c(2.12,3.33,4.90,6.22,7.01,7.51,7.87,8.44,8.76,8.98)
data <- cbind(s,v)
data<-data.frame(data)
WAHKineticEnzyme(substrate=data$s,velocity=data$v,removeoutliers=TRUE,deepening=TRUE)
}
